+++ tests/file_tests.cpp
@@ -16,6 +16,7 @@ TEST_CASE("read_rect_16bit") {
   REQUIRE(bw64File->sampleRate() == 44100u);
   REQUIRE(bw64File->channels() == 2u);
   REQUIRE(bw64File->numberOfFrames() == 22050u);
+  bw64File->close();
 }
 
 TEST_CASE("read_rect_24bit") {
@@ -25,6 +26,7 @@ TEST_CASE("read_rect_24bit") {
   REQUIRE(bw64File->sampleRate() == 44100u);
   REQUIRE(bw64File->channels() == 2u);
   REQUIRE(bw64File->numberOfFrames() == 22050u);
+  bw64File->close();
 }
 
 TEST_CASE("read_rect_32bit") {
@@ -34,6 +36,7 @@ TEST_CASE("read_rect_32bit") {
   REQUIRE(bw64File->sampleRate() == 44100u);
   REQUIRE(bw64File->channels() == 2u);
   REQUIRE(bw64File->numberOfFrames() == 22050u);
+  bw64File->close();
 }
 
 TEST_CASE("read_rect_24bit_rf64") {
@@ -43,6 +46,7 @@ TEST_CASE("read_rect_24bit_rf64") {
   REQUIRE(bw64File->sampleRate() == 44100u);
   REQUIRE(bw64File->channels() == 2u);
   REQUIRE(bw64File->numberOfFrames() == 22050u);
+  bw64File->close();
 }
 
 TEST_CASE("read_rect_24bit_noriff") {
@@ -68,6 +72,7 @@ TEST_CASE("read_noise_24bit_uneven_data_chunk_size") {
   REQUIRE(bw64File->chnaChunk() != nullptr);
   REQUIRE(bw64File->hasChunk(utils::fourCC("axml")) == false);
   REQUIRE(bw64File->axmlChunk() == nullptr);
+  bw64File->close();
 }
 
 TEST_CASE("read_check_chunks") {
@@ -78,6 +83,7 @@ TEST_CASE("read_check_chunks") {
   REQUIRE(bw64File->hasChunk(utils::fourCC("fmt ")) == true);
   REQUIRE(bw64File->hasChunk(utils::fourCC("chna")) == false);
   REQUIRE(bw64File->hasChunk(utils::fourCC("axml")) == false);
+  bw64File->close();
 }
 
 TEST_CASE("read_seek_tell") {
@@ -108,6 +114,8 @@ TEST_CASE("read_seek_tell") {
   REQUIRE(bw64File->tell() == 11025);
   bw64File->seek(INT32_MAX, std::ios::end);
   REQUIRE(bw64File->tell() == 22050);
+
+  bw64File->close();
 }
 
 TEST_CASE("write_16bit") {
@@ -121,6 +129,7 @@ TEST_CASE("write_16bit") {
   std::vector<float> data(frames * bw64File->channels(), 0.0);
   bw64File->write(&data[0], frames);
   REQUIRE(bw64File->framesWritten() == frames);
+  bw64File->close();
 }
 
 TEST_CASE("write_24bit") {
@@ -134,6 +143,7 @@ TEST_CASE("write_24bit") {
   std::vector<float> data(frames * bw64File->channels(), 0.0);
   bw64File->write(&data[0], frames);
   REQUIRE(bw64File->framesWritten() == frames);
+  bw64File->close();
 }
 
 TEST_CASE("write_32bit") {
@@ -147,6 +157,7 @@ TEST_CASE("write_32bit") {
   std::vector<float> data(frames * bw64File->channels(), 0.0);
   bw64File->write(&data[0], frames);
   REQUIRE(bw64File->framesWritten() == frames);
+  bw64File->close();
 }
 
 void writeClipped(const std::string& filename, uint16_t bitDepth,
@@ -155,6 +166,7 @@ void writeClipped(const std::string& filename, uint16_t bitDepth,
   auto bw64File = writeFile(filename, channels, sampleRate, bitDepth);
   std::vector<float> data(frames * channels, 2.f);
   bw64File->write(&data[0], frames);
+  bw64File->close();
 }
 
 TEST_CASE("write_read_16bit_clipped") {
@@ -218,6 +230,7 @@ void writeRandom(const std::string& filename, uint16_t bitDepth,
 
   auto bw64File = writeFile(filename, channels, sampleRate, bitDepth);
   bw64File->write(&data[0], frames);
+  bw64File->close();
 }
 
 TEST_CASE("write_read_riff_header") {
@@ -228,7 +241,7 @@ TEST_CASE("write_read_riff_header") {
   /**
    * 'WAVE' chunk header  :      4 bytes
    * 'JUNK' chunk header  :      8 bytes
-   * 'JUNK' chunk payload :     28 bytes
+   * 'JUNK' chunk payload :     40 bytes
    * 'fmt ' chunk header  :      8 bytes
    * 'fmt ' chunk payload :     16 bytes
    * 'chna' chunk header  :      8 bytes
@@ -236,9 +249,9 @@ TEST_CASE("write_read_riff_header") {
    * (numTracks/numUIDs) 'data' chunk header  :      8 bytes 'data' chunk
    * payload : 19.200 bytes (2 channels * 2 bytes per samples * 4800 frames)
    * --------------------------------------------------------------------------
-   *                        60244 bytes
+   *                        60256 bytes
    */
-  REQUIRE(bw64File->fileSize() == 60244);
+  REQUIRE(bw64File->fileSize() == 60256);
 }
 
 TEST_CASE("write_read_96000") {
@@ -257,6 +270,26 @@ TEST_CASE("can_read_all_frames") {
   REQUIRE(readSampleCount == 13);
 }
 
+TEST_CASE("write_read_odd_chunks_after") {
+  int frames = 13;
+
+  {
+    std::vector<float> data(frames, 0.5);
+    auto writer = writeFile("write_read_odd_chunks_after.wav", 1, 48000, 24);
+    writer->setAxmlChunk(std::make_shared<AxmlChunk>("axml"));
+    writer->write(&data[0], frames);
+    writer->close();
+  }
+
+  {
+    auto reader = readFile("write_read_odd_chunks_after.wav");
+    REQUIRE(reader->numberOfFrames() == frames);
+    auto axml = reader->axmlChunk();
+    REQUIRE(axml);
+    REQUIRE(axml->data() == "axml");
+  }
+}
+
 TEST_CASE("write_read_big", "[.big]") {
   uint64_t frames = 0x90000000UL;
   uint64_t blockSize = 0x1000UL;
@@ -271,6 +304,8 @@ TEST_CASE("write_read_big", "[.big]") {
 
     for (uint64_t frame = 0; frame < frames; frame += block.size())
       bw64File->write(&block[0], block.size());
+
+    bw64File->close();
   }
 
   {
@@ -289,4 +324,94 @@ TEST_CASE("write_read_big", "[.big]") {
       }
     }
   }
+
+  remove(filename.c_str());
+}
+
+TEST_CASE("write_read_big_axml", "[.big]") {
+  std::string filename = "big_axml.wav";
+  size_t blockSize = 1000;
+  uint64_t axml_size = 0x100000000ull;
+
+  const char* pattern = "AXML";
+
+  {
+    std::string axml_data(axml_size, 0);
+    for (size_t i = 0; i < axml_data.size(); i++) axml_data[i] = pattern[i % 4];
+
+    auto axml = std::make_shared<AxmlChunk>(std::move(axml_data));
+
+    auto bw64File = writeFile(filename, 1, 48000, 16, nullptr, axml);
+
+    std::vector<float> block(blockSize);
+    for (size_t i = 0; i < block.size(); i++)
+      block[i] = (i % 2 == 0) ? 0.5f : 0.0f;
+    bw64File->write(&block[0], block.size());
+
+    bw64File->close();
+  }
+
+  {
+    auto bw64File = readFile(filename);
+
+    // check samples
+    REQUIRE(bw64File->numberOfFrames() == blockSize);
+    std::vector<float> block(blockSize);
+    auto readFrames = bw64File->read(&block[0], blockSize);
+    REQUIRE(readFrames == blockSize);
+    for (size_t i = 0; i < blockSize; i++) {
+      if (i % 2 == 0)
+        REQUIRE(block[i] == Approx(0.5f).epsilon(1e-2));
+      else
+        REQUIRE(block[i] == Approx(0.0f).epsilon(1e-2));
+    }
+
+    // check axml
+    auto axml = bw64File->axmlChunk();
+    REQUIRE(axml);
+
+    auto& axml_data = axml->data();
+    REQUIRE(axml_data.size() == axml_size);
+
+    for (size_t i = 0; i < axml_data.size(); i++) {
+      REQUIRE(axml_data[i] == pattern[i % 4]);
+    }
+  }
+
+  remove(filename.c_str());
+}
+
+/// dummy chunk that just writes spaces
+class MegaChunk : public Chunk {
+ public:
+  MegaChunk(uint32_t id, uint64_t size) : id_(id), size_(size) {}
+
+  uint32_t id() const override { return id_; }
+  uint64_t size() const override { return size_; }
+
+  void write(std::ostream& stream) const override {
+    for (size_t i = 0; i < size_; i++) stream << ' ';
+  }
+
+ private:
+  uint32_t id_;
+  uint64_t size_;
+};
+
+TEST_CASE("write_too_many_big_chunks", "[.big]") {
+  std::string filename = "too_many_big_chunks.wav";
+
+  auto bw64File = writeFile(filename, 1, 48000, 16, nullptr);
+
+  // TODO: setAxmlChunk should really be renamed...
+  bw64File->setAxmlChunk(
+      std::make_shared<MegaChunk>(utils::fourCC("mega"), 0x100000000ull));
+  bw64File->setAxmlChunk(
+      std::make_shared<MegaChunk>(utils::fourCC("megb"), 0x100000000ull));
+
+  REQUIRE_THROWS_WITH(
+      bw64File->close(),
+      "ds64 chunk is too large (52 bytes) to overwrite JUNK chunk (40 bytes)");
+
+  remove(filename.c_str());
 }